Rename NoteItem to FeedNoteItem
