+++ gratitude-notes-core-skeleton/src/components/FeedNoteItem.tsx
